pyproject.toml
[tool.poetry]
name = "todoappmysql"
version = "0.1.0"
description = ""
authors = ["mukhtar-ahmed <kh.mukhtar60@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
SQLAlchemy = "^2.0.32"
pydantic = "^2.8.2"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest = "^8.3.2"
httpx = "^0.27.2"
python-multipart = "^0.0.9"
bcrypt = "4.0.1"
pytest-asyncio = "^0.24.0"
aiofiles = "^24.1.0"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
